chore: Move URI parsing logic to uniformResourceIdentifier.go
