--- authorization/v2/parser_test.go
@@ -1,1 +1,0 @@
-package v2
